# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Indigo API Documentation
  description: "This is an API Documentation for the Indigo mibile and web application."
  contact:
    name: Nikethan
    url: https://www.linkedin.com/in/nikethan-selvanathan-a81928101
    email: nikesky273@gmail.com

# during dev, should point to your local machine
host: 127.0.0.1:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /organization/{ID}:
    x-swagger-router-controller: OrganizationDocumentController
    get:
      description: Get Organization by ID
      summary: Get Organization by ID
      tags:
        - Organizations
      # used as the method name of the controller
      operationId: getOrganizationByID
      parameters:
        - name: ID
          in: path
          description: ID of Organization
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OrganizationsDef"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /organization:
    x-swagger-router-controller: OrganizationDocumentController
    get:
      description: Retrieves all Organizations documents from the DB
      summary: Get all the Organizations documents of APIs
      tags:
        - Organizations
      # used as the method name of the controller
      operationId: getAllOrganizations
      parameters:
        - name: limit
          in: query
          description: No of organization to return
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OrganizationsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a Organization document
      summary: Add a new Organization Document in DB
      tags: 
        - Organizations
      # used as the method name of the controller
      operationId: createOrganization
      parameters:
        - name: request
          in: body
          description: Request to the endpoint
          required: true
          schema:
            $ref: "#/definitions/OrganizationDef"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a Organization document
      summary: Update a Organization Document in DB
      tags: 
        - Organizations
      # used as the method name of the controller
      operationId: updateOrganization
      parameters:
        - name: organizationID
          in: query
          description: Id of the organization Document
          required: true
          type: string
        - name: request
          in: body
          description: Request to the endpoint
          required: true
          schema:
            $ref: "#/definitions/OrganizationDef"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete a Organization document
      summary: Delete a Organization Document from DB
      tags: 
        - Organizations
      # used as the method name of the controller
      operationId: deleteOrganization
      parameters:
        - name: organizationID
          in: query
          description: Id of the organization Document
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  # Branch API's
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    required:
      - id
    properties:
      id:
        type: string
...
  OrganizationsResponse:
    required:
      - organizations
    properties:
      organizations:
        type: array
        items:
          $ref: "#/definitions/OrganizationsDef"
...
  OrganizationsDef:
    properties:
      _id:
        type: string
      _rev:
        type: string
      name:
        type: string
      image:
        type: string
      founder:
        type: string
      founded:
        type: string
      description:
        type: string
      collection:
        type: string
  OrganizationDef:
    properties:
      name:
        type: string
      image:
        type: string
      founder:
        type: string
      founded:
        type: string
      description:
        type: string
...